chore: a standalone keys implementation for the legacy library
