+++ packages/legacy/keys/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+    "env": {
+        "mocha": true
+    }
+}
