+++ packages/legacy/keys/.releaserc.json
@@ -1,0 +1,8 @@
+{
+    "commitPaths": ["../"],
+    "repositoryUrl": "git@github.com:solana-labs/solana-web3.js.git",
+    "preset": "conventionalcommits",
+    "presetConfig": {
+        "issueUrlFormat": "{{host}}/{{owner}}/solana-web3.js/issues/{{id}}"
+    }
+}
