+++ packages/legacy/keys/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "@solana/web3.js-legacy-sham",
-    "version": "1.0.0-development",
-    "description": "A drop-in replacement for @solana/web3.js@1 that you can use to break your dependency on the legacy library, particularly when one of your dependencies pulls it in transitively",
+    "name": "@solana/keys-legacy",
+    "version": "0.0.0-development",
+    "description": "Helpers for generating and transforming key material",
     "exports": {
         "browser": {
             "import": "./dist/index.browser.js",
@@ -34,19 +34,14 @@
     ],
     "scripts": {
         "compile:js": "tsup --config build-scripts/tsup.config.package.ts",
-        "compile:typedefs": "tsc -p ./tsconfig.declarations.json && node node_modules/build-scripts/add-js-extension-to-types.mjs",
+        "compile:typedefs": "tsc -p ./tsconfig.declarations.json",
         "dev": "jest -c node_modules/test-config/jest-dev.config.ts --rootDir . --watch",
-        "prepublishOnly": "version-from-git --no-git-tag-version --template experimental.short",
-        "publish-packages": "pnpm publish --tag experimental --access public --no-git-checks",
         "style:fix": "pnpm eslint --fix src/* && pnpm prettier -w src/* package.json",
         "test:lint": "jest -c node_modules/test-config/jest-lint.config.ts --rootDir . --silent",
+        "test:lint:fix": "eslint src/ test/ --fix --ext .js,.ts",
         "test:prettier": "jest -c node_modules/test-config/jest-prettier.config.ts --rootDir . --silent",
-        "test:treeshakability:browser": "agadoo dist/index.browser.js",
-        "test:treeshakability:native": "agadoo dist/index.native.js",
-        "test:treeshakability:node": "agadoo dist/index.node.js",
         "test:typecheck": "tsc --noEmit",
-        "test:unit:browser": "jest -c node_modules/test-config/jest-unit.config.browser.ts --rootDir . --silent",
-        "test:unit:node": "jest -c node_modules/test-config/jest-unit.config.node.ts --rootDir . --silent"
+        "test:unit:node": "cross-env NODE_ENV=test TS_NODE_COMPILER_OPTIONS='{ \"module\": \"commonjs\", \"target\": \"es2019\" }' ts-mocha --require esm './test/**/*.test.ts'"
     },
     "author": "Solana Labs Maintainers <maintainers@solanalabs.com>",
     "license": "MIT",
@@ -65,35 +60,45 @@
         "node": ">=17.4"
     },
     "dependencies": {
-        "@noble/ed25519": "^2.0.0",
         "@noble/hashes": "^1.3.2",
-        "@solana/addresses": "workspace:*",
-        "@solana/transactions": "workspace:*"
+        "bn.js": "^5.2.1",
+        "borsh": "^0.7.0",
+        "bs58": "^4.0.1",
+        "buffer": "6.0.3",
+        "legacy-shared": "workspace:*"
     },
     "devDependencies": {
         "@solana/eslint-config-solana": "^1.0.2",
-        "@solana/web3.js-legacy": "workspace:../legacy/library",
         "@swc/jest": "^0.2.29",
         "@types/bn.js": "^5.1.1",
-        "@types/jest": "^29.5.11",
-        "@types/node": "18.11.19",
-        "@typescript-eslint/eslint-plugin": "^6.13.2",
+        "@types/bs58": "^4.0.1",
+        "@types/chai": "^4.3.5",
+        "@types/chai-as-promised": "^7.1.8",
+        "@types/mocha": "^10.0.6",
+        "@types/sinon-chai": "^3.2.9",
+        "@typescript-eslint/eslint-plugin": "^6.7.0",
         "@typescript-eslint/parser": "^6.3.0",
-        "agadoo": "^3.0.0",
         "build-scripts": "workspace:*",
-        "eslint-plugin-jest": "^27.4.2",
-        "eslint-plugin-sort-keys-fix": "^1.1.2",
+        "chai": "^4.3.7",
+        "chai-as-promised": "^7.1.1",
+        "cross-env": "7.0.3",
         "eslint": "^8.45.0",
-        "jest-environment-jsdom": "^29.7.0",
+        "eslint-plugin-jest": "^27.4.2",
+        "eslint-plugin-mocha": "^10.1.0",
+        "eslint-plugin-sort-keys-fix": "^1.1.2",
+        "jest": "^29.7.0",
         "jest-runner-eslint": "^2.1.2",
         "jest-runner-prettier": "^1.0.0",
-        "jest": "^29.7.0",
+        "mocha": "^10.2.0",
         "prettier": "^3.1",
+        "semantic-release-plus": "^20.0.0",
+        "sinon": "^15.2.0",
+        "sinon-chai": "^3.7.0",
         "test-config": "workspace:*",
+        "ts-mocha": "^10.0.0",
         "tsconfig": "workspace:*",
         "tsup": "^8.0.1",
-        "typescript": "^5.2.2",
-        "version-from-git": "^1.1.1"
+        "typescript": "^5.2.2"
     },
     "bundlewatch": {
         "defaultCompression": "gzip",
