+++ packages/legacy/keys/src/index.ts
@@ -1,0 +1,6 @@
+import { Enum, SOLANA_SCHEMA, Struct } from './borsh-schema';
+export const Enum_DO_NOT_USE_OR_YOU_WILL_BE_FIRED = Enum;
+export const SOLANA_SCHEMA_DO_NOT_USE_OR_YOU_WILL_BE_FIRED = SOLANA_SCHEMA;
+export const Struct_DO_NOT_USE_OR_YOU_WILL_BE_FIRED = Struct;
+export * from './key-pair';
+export * from './public-key';
