+++ packages/legacy/keys/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "$schema": "https://json.schemastore.org/tsconfig",
+    "compilerOptions": {
+        "esModuleInterop": true,
+        "lib": ["DOM", "ES2015"]
+    },
+    "display": "@solana/keys (legacy)",
+    "extends": "tsconfig/base.json",
+    "include": ["src", "test"]
+}
