+++ packages/legacy/library/package.json
@@ -57,6 +57,7 @@
     "@noble/curves": "^1.2.0",
     "@noble/hashes": "^1.3.2",
     "@solana/buffer-layout": "^4.0.1",
+    "@solana/keys": "workspace:@solana/keys-legacy@*",
     "agentkeepalive": "^4.5.0",
     "bigint-buffer": "^1.1.5",
     "bn.js": "^5.2.1",
