+++ packages/legacy/library/src/bpf-loader-deprecated.ts
@@ -1,4 +1,4 @@
-import {PublicKey} from './publickey';
+import {PublicKey} from '@solana/keys';
 
 export const BPF_LOADER_DEPRECATED_PROGRAM_ID = new PublicKey(
   'BPFLoader1111111111111111111111111111111111',
