+++ packages/legacy/library/src/programs/address-lookup-table/state.ts
@@ -1,8 +1,8 @@
 import * as BufferLayout from '@solana/buffer-layout';
+import {PublicKey} from '@solana/keys';
 
 import assert from '../../utils/assert';
 import * as Layout from '../../layout';
-import {PublicKey} from '../../publickey';
 import {u64} from '../../utils/bigint';
 import {decodeData} from '../../account-data';
