+++ packages/legacy/library/src/programs/system.ts
@@ -1,4 +1,5 @@
 import * as BufferLayout from '@solana/buffer-layout';
+import {PublicKey} from '@solana/keys';
 
 import {
   encodeData,
@@ -8,7 +9,6 @@ import {
 } from '../instruction';
 import * as Layout from '../layout';
 import {NONCE_ACCOUNT_LENGTH} from '../nonce-account';
-import {PublicKey} from '../publickey';
 import {SYSVAR_RECENT_BLOCKHASHES_PUBKEY, SYSVAR_RENT_PUBKEY} from '../sysvar';
 import {Transaction, TransactionInstruction} from '../transaction';
 import {u64} from '../utils/bigint';
