+++ packages/legacy/library/src/programs/vote.ts
@@ -1,4 +1,5 @@
 import * as BufferLayout from '@solana/buffer-layout';
+import {PublicKey} from '@solana/keys';
 
 import {
   encodeData,
@@ -7,7 +8,6 @@ import {
   IInstructionInputData,
 } from '../instruction';
 import * as Layout from '../layout';
-import {PublicKey} from '../publickey';
 import {SystemProgram} from './system';
 import {SYSVAR_CLOCK_PUBKEY, SYSVAR_RENT_PUBKEY} from '../sysvar';
 import {Transaction, TransactionInstruction} from '../transaction';
