+++ packages/legacy/library/src/transaction/legacy.ts
@@ -1,3 +1,5 @@
+import {PublicKey} from '@solana/keys';
+import type {Signer} from '@solana/keys';
 import bs58 from 'bs58';
 import {Buffer} from 'buffer';
 import {sign, toBuffer, verify} from 'legacy-shared';
@@ -5,10 +7,8 @@ import {sign, toBuffer, verify} from 'legacy-shared';
 import {PACKET_DATA_SIZE, SIGNATURE_LENGTH_IN_BYTES} from './constants';
 import {Connection} from '../connection';
 import {Message} from '../message';
-import {PublicKey} from '../publickey';
 import * as shortvec from '../utils/shortvec-encoding';
 import invariant from '../utils/assert';
-import type {Signer} from '../keypair';
 import type {Blockhash} from '../blockhash';
 import type {CompiledInstruction} from '../message';
