+++ packages/legacy/library/src/transaction/message.ts
@@ -1,8 +1,9 @@
+import {PublicKey} from '@solana/keys';
+
 import {AccountKeysFromLookups} from '../message/account-keys';
 import assert from '../utils/assert';
 import {Blockhash} from '../blockhash';
 import {Message, MessageV0, VersionedMessage} from '../message';
-import {PublicKey} from '../publickey';
 import {AddressLookupTableAccount} from '../programs';
 import {AccountMeta, TransactionInstruction} from './legacy';
 import {toBuffer} from 'legacy-shared';
