+++ packages/legacy/library/src/transaction/versioned.ts
@@ -1,13 +1,12 @@
 import * as BufferLayout from '@solana/buffer-layout';
+import {Signer, PublicKey} from '@solana/keys';
 import {sign} from 'legacy-shared';
 
-import {Signer} from '../keypair';
 import assert from '../utils/assert';
 import {VersionedMessage} from '../message/versioned';
 import {SIGNATURE_LENGTH_IN_BYTES} from './constants';
 import * as shortvec from '../utils/shortvec-encoding';
 import * as Layout from '../layout';
-import {PublicKey} from '../publickey';
 
 export type TransactionVersion = 'legacy' | 0;
