+++ packages/legacy/library/src/utils/index.ts
@@ -1,4 +1,3 @@
-export * from './borsh-schema';
 export * from './cluster';
 export * from './send-and-confirm-raw-transaction';
 export * from './send-and-confirm-transaction';
