+++ packages/legacy/library/test/message-tests/v0.test.ts
@@ -1,6 +1,7 @@
 import bs58 from 'bs58';
 import {expect} from 'chai';
 import {sha256} from '@noble/hashes/sha256';
+import {PublicKey} from '@solana/keys';
 
 import {
   MessageAccountKeys,
@@ -8,7 +9,6 @@ import {
   MessageV0,
 } from '../../src/message';
 import {TransactionInstruction} from '../../src/transaction';
-import {PublicKey} from '../../src/publickey';
 import {AddressLookupTableAccount} from '../../src/programs';
 
 function createTestKeys(count: number): Array<PublicKey> {
